requirements.txt
requestsbeautifulsoup4